backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
Pillow==10.1.0
face-recognition==1.3.0
opencv-python==4.8.1.78
python-multipart==0.0.6
djangorestframework-simplejwt==5.3.0
python-decouple==3.8